refactor: extract residentService from checkoutService
Move all building, unit, and resident API functions into a dedicated
residentService. checkoutService now only contains checkout processing
functions (processWelcomeBasket, processGeneralItems, checkPastCheckout).

Also adds cross-building resident search using client-side filtering
on a full residents load via the ResidentsByBuilding view.
